+++ cmd/bounce/main.go
@@ -11,21 +11,33 @@ import (
 	"errors"
 	"flag"
 	"fmt"
-	"log"
 	"net/http"
 	"os"
-	"time"
 
 	"github.com/atc0005/bounce/config"
 	"github.com/atc0005/bounce/routes"
+
+	"github.com/apex/log"
+	"github.com/apex/log/handlers/cli"
+	"github.com/apex/log/handlers/discard"
+	"github.com/apex/log/handlers/json"
+	"github.com/apex/log/handlers/logfmt"
+	"github.com/apex/log/handlers/text"
 )
+
+// for handler in cli discard es graylog json kinesis level logfmt memory
+// multi papertrail text delta; do go get
+// github.com/apex/log/handlers/${handler}; done
 
 // see templates.go for the hard-coded HTML/CSS template used for the index
 // page
 
 func main() {
 
-	log.Println("DEBUG: Initializing application")
+	// This will use default logging settings (level filter, destination)
+	// as the application hasn't "booted up" far enough to apply custom
+	// choices yet.
+	log.Debug("Initializing application")
 
 	appConfig, err := config.NewConfig()
 	if err != nil {
@@ -35,7 +47,41 @@ func main() {
 		log.Fatalf("Failed to initialize application: %s", err)
 	}
 
-	log.Printf("DEBUG: %+v\n", appConfig)
+	var logOutput *os.File
+	switch appConfig.LogOutput {
+	case config.LogOutputStderr:
+		logOutput = os.Stderr
+	case config.LogOutputStdout:
+		logOutput = os.Stdout
+	}
+
+	switch appConfig.LogFormat {
+	case config.LogFormatCLI:
+		log.SetHandler(cli.New(logOutput))
+	case config.LogFormatJSON:
+		log.SetHandler(json.New(logOutput))
+	case config.LogFormatLogFmt:
+		log.SetHandler(logfmt.New(logOutput))
+	case config.LogFormatText:
+		log.SetHandler(text.New(logOutput))
+	case config.LogFormatDiscard:
+		log.SetHandler(discard.New())
+	}
+
+	switch appConfig.LogLevel {
+	case config.LogLevelFatal:
+		log.SetLevel(log.FatalLevel)
+	case config.LogLevelError:
+		log.SetLevel(log.ErrorLevel)
+	case config.LogLevelWarn:
+		log.SetLevel(log.WarnLevel)
+	case config.LogLevelInfo:
+		log.SetLevel(log.InfoLevel)
+	case config.LogLevelDebug:
+		log.SetLevel(log.DebugLevel)
+	}
+
+	log.Debugf("AppConfig: %+v", appConfig)
 
 	// SETUP ROUTES
 	// See handlers.go for handler definitions
@@ -74,14 +120,19 @@ func main() {
 	// small-to-medium on-premise API (e.g., not over a public Internet link
 	// where clients are expected to be slow)
 	httpServer := &http.Server{
-		ReadHeaderTimeout: 20 * time.Second,
+		ReadHeaderTimeout: config.HTTPServerReadHeaderTimeout,
+		ReadTimeout:       config.HTTPServerReadTimeout,
+		WriteTimeout:      config.HTTPServerWriteTimeout,
 		Handler:           mux,
 		Addr:              fmt.Sprintf("%s:%d", appConfig.LocalIPAddress, appConfig.LocalTCPPort),
 	}
 
 	// listen on specified port on ALL IP Addresses, block until app is terminated
-	log.Printf("Listening on %s port %d ",
+	log.Infof("Listening on %s port %d ",
 		appConfig.LocalIPAddress, appConfig.LocalTCPPort)
 
-	log.Fatal(httpServer.ListenAndServe())
+	// TODO: This can be handled in a cleaner fashion?
+	if err := httpServer.ListenAndServe(); err != nil {
+		log.Fatal(err.Error())
+	}
 }